requirements.txt
alembic==1.12.0
fastapi==0.103.2
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
SQLAlchemy==2.0.22
uvicorn==0.23.2
pytest==7.4.2
argon2-cffi==23.1.0
PyJWT==2.8.0